filesio: Handle multiple file extension in output definition parser
`parse_output_definition` only returns the default file extension even
if multiple formats are supported.

This issue is fixed by parsing the `fileFilter` of the file
definition. This allows to retrieve all the supported
extensions. Besides, the default file format is moved at the beginning
of the `supported_formats` list to ensure that it is interpreted as
the default extension.